requirements.txt
# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# OpenAI API
openai==0.28.1

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.3

# Visualization
plotly==5.17.0

# Table formatting and export
tabulate==0.9.0
openpyxl==3.1.2

# Progress bar
tqdm==4.66.1
